Flatten away "Private" subdirs to allow including individual headers
