+++ Sources/Plasma/NucleusLib/pnAsyncCore/CMakeLists.txt
@@ -1,20 +1,20 @@
 set(pnAsyncCore_HEADERS
     Pch.h
+    pnAcCore.h
+    pnAcIo.h
+    pnAcLog.h
+    pnAcThread.h
+    pnAcTimer.h
     pnAsyncCore.h
 )
 
-set(pnAsyncCore_PRIVATE
-    Private/pnAcCore.h
-    Private/pnAcIo.h
-    Private/pnAcLog.cpp
-    Private/pnAcLog.h
-    Private/pnAcThread.h
-    Private/pnAcTimer.h
+set(pnAsyncCore_SOURCES
+    pnAcLog.cpp
 )
 
 plasma_library(pnAsyncCore
     FOLDER NucleusLib
-    SOURCES ${pnAsyncCore_PRIVATE} ${pnAsyncCore_HEADERS}
+    SOURCES ${pnAsyncCore_HEADERS} ${pnAsyncCore_SOURCES}
 )
 
 target_include_directories(pnAsyncCore PRIVATE "${PLASMA_SOURCE_ROOT}/PubUtilLib")
@@ -31,4 +31,4 @@ target_link_libraries(
 )
 
 source_group("Header Files" FILES ${pnAsyncCore_HEADERS})
-source_group("Private" FILES ${pnAsyncCore_PRIVATE})
+source_group("Source Files" FILES ${pnAsyncCore_SOURCES})
