+++ Sources/Plasma/PubUtilLib/plNetGameLib/plNglAuth.cpp
@@ -40,7 +40,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 *==LICENSE==*/
 
-#include "../Pch.h"
+#include "Pch.h"
 
 #include <regex>
 #include "pnEncryption/plChallengeHash.h"
