+++ Sources/Plasma/PubUtilLib/plNetGameLib/plNglCore.cpp
@@ -40,7 +40,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 *==LICENSE==*/
 
-#include "../Pch.h"
+#include "Pch.h"
 
 
 namespace Ngl {
