+++ Sources/Plasma/PubUtilLib/plNetGameLib/plNglFile.cpp
@@ -40,7 +40,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 *==LICENSE==*/
 
-#include "../Pch.h"
+#include "Pch.h"
 
 #include "pnUtils/pnUtStr.h"
