+++ Sources/Plasma/PubUtilLib/plNetGameLib/plNglGame.cpp
@@ -40,7 +40,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 *==LICENSE==*/
 
-#include "../Pch.h"
+#include "Pch.h"
 
 namespace Ngl { namespace Game {
 /*****************************************************************************
